chatbot-api/requirements.txt
fastapi
uvicorn
requests
google-genai
python-dotenv